RagService/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
chromadb>=0.5.7
sentence-transformers==3.0.0
groq==0.4.2
mysql-connector-python==8.4.0
python-multipart>=0.0.7
